Rewrite code to handle git bug
Change how git is called within the script to handle calling
git in a subdirectory. Previous calls to git set the --git-dir
directory to game-repo/.git and the --work-tree to game-repo/

This worked fine until a command did not work as expected:
git reset --hard. It failed to reset the repository while
reporting that it had. Probably a bug in git.

Upon further examination, the -C option, which specifies a
normal git repository in the provided subfolder works as expected.
My 2.17 man pages say the two options are equivalent. I have made
so man calls to subprocess that it will be a pain to switch
them all over.

TODO: confirm the git reset --hard bug in my test suite.