Add support for additional diff types.
Introduce a function for diffing all changes since previous
commit, and a function for diffing all changes between branches.
Bring the total change inspection capacity for the game to
four, as opposed to the highly flexible and hence more confusing
five variations of git diff which implements the entire refspec.