Standardize git calls
Due to confusing behavior surrounding calling git on subdirectories,
make all calls to git in the form of
    G.GIT, '-C', repodir
where G.GIT is the globally defined command to the git binary,
and repodir is an appropriately scoped global variable (global
to the class or global to the test suite). This will ensure git
behaves properly (particularly when git reset --hard is called)
or can be easily modified if changes are to be made.