Fix when program recognizes player death.
Alter the program that the main logic will recognize all changes
of state in living / death. Have gitcli keep track of one piece
of game data (living/death status), and have every method that
could change the alive.json file return True if it does so.

It should be noted that returning True to a do_* method ends
the loop so any file changing git commands from the git cli must
return True if the character is alive afterwards. However, within
the game cli those same commands must return True if the character
is dead afterwards.

Personal Note: I've decided to refer to the possibility of killing
yourself using git commands as bug #1, because I did not originally
anticipate it was possible.