Track player's alive status in git cli.
Make an exception for separation of logic between the game and
git cli to track one player attribute, whether the character is
alive. The git cli needs this information in the main game loop,
so include it. If too many exceptions like this start being made,
consider creating a parent class for the git cli that only has
git logic in it.